added more javadoc for the new stuff I have added
