+++ src/main/java/com/github/onsdigital/brian/handlers/responses/JsonTransformer.java
@@ -3,11 +3,18 @@ package com.github.onsdigital.brian.handlers.responses;
 import com.google.gson.Gson;
 import spark.ResponseTransformer;
 
+/**
+ * Implementation of the Spark {@link JsonTransformer}. Used by the Spark framework to convert response objects to JSON.
+ */
 public class JsonTransformer implements ResponseTransformer {
+
     private static final JsonTransformer instance = new JsonTransformer();
 
     private Gson gson;
 
+    /**
+     * Singleton instance - use {@link #getInstance()} to obtain it.
+     */
     private JsonTransformer() {
         this.gson = new Gson();
     }
@@ -17,6 +24,9 @@ public class JsonTransformer implements ResponseTransformer {
         return gson.toJson(o);
     }
 
+    /**
+     * @return a singleton instance of the Transformer.
+     */
     public static JsonTransformer getInstance() {
         return instance;
     }
