+++ src/main/java/com/github/onsdigital/brian/handlers/responses/Message.java
@@ -1,9 +1,17 @@
 package com.github.onsdigital.brian.handlers.responses;
 
+/**
+ * POJO used for simple JSON responses.
+ */
 public class Message {
 
     private String mesage;
 
+    /**
+     * Costruct a new Message
+     *
+     * @param mesage the value of the message field you wish to set.
+     */
     public Message(String mesage) {
         this.mesage = mesage;
     }
