+++ src/main/java/com/github/onsdigital/brian/readers/csdb/DataSetReaderCSDB.java
@@ -13,7 +13,7 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.regex.Pattern;
 
-import static com.github.onsdigital.brian.logging.Logger.logEvent;
+import static com.github.onsdigital.brian.logging.LogEvent.logEvent;
 
 /**
  * Created by thomasridd on 10/03/15.
